+++ playground/react-native-playground/CHANGELOG.md
@@ -7,6 +7,45 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0]
+
+### Changed
+
+- Update wagmi `metamask-connector` to unwrap `.signature` / `.result` from the new `connectAndSign` / `connectWith` return values ([#266](https://github.com/MetaMask/connect-monorepo/pull/266))
+- Bump workspace dependencies:
+  - @metamask/connect-evm@1.0.0
+  - @metamask/connect-solana@1.0.0
+
+## [0.3.6]
+
+### Changed
+
+- Bump workspace dependencies:
+  - @metamask/connect-evm@0.11.2
+  - @metamask/connect-multichain@0.12.1
+
+## [0.3.5]
+
+### Changed
+
+- Bump workspace dependencies:
+  - @metamask/connect-evm@0.11.1
+
+## [0.3.4]
+
+### Changed
+
+- Bump workspace dependencies:
+  - @metamask/connect-evm@0.11.0
+
+## [0.3.3]
+
+### Changed
+
+- Bump workspace dependencies:
+  - @metamask/connect-evm@0.10.0
+  - @metamask/connect-multichain@0.12.0
+
 ## [0.3.2]
 
 ### Changed
@@ -98,7 +137,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.3.2...HEAD
+[Unreleased]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.4.0...HEAD
+[0.4.0]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.3.6...@metamask/react-native-playground@0.4.0
+[0.3.6]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.3.5...@metamask/react-native-playground@0.3.6
+[0.3.5]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.3.4...@metamask/react-native-playground@0.3.5
+[0.3.4]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.3.3...@metamask/react-native-playground@0.3.4
+[0.3.3]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.3.2...@metamask/react-native-playground@0.3.3
 [0.3.2]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.3.1...@metamask/react-native-playground@0.3.2
 [0.3.1]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.3.0...@metamask/react-native-playground@0.3.1
 [0.3.0]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.2.1...@metamask/react-native-playground@0.3.0